requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
backports.tarfile==1.1.1
beautifultable==1.1.0
cachetools==5.3.3
certifi==2024.2.2
cex-adaptors==1.0.3
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dex-adaptors==0.0.3
distlib==0.3.8
dnspython==2.6.1
docutils==0.21.2
exceptiongroup==1.2.1
fastapi==0.110.0
filelock==3.14.0
fonttools==4.50.0
frozenlist==1.4.1
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
gql==3.5.0
graphql-core==3.2.3
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
identify==2.5.36
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
keyring==25.2.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
matplotlib==3.8.3
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.5
nh3==0.2.17
nodeenv==1.8.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
pillow==10.2.0
pkginfo==1.10.0
platformdirs==4.2.1
pluggy==1.5.0
pre-commit==3.7.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.18.0
pygsheets==2.0.6
pymongo==4.6.3
pyparsing==3.1.2
pytest==8.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.0.1
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
readme_renderer==43.0
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
rsa==4.9
six==1.16.0
slack_sdk==3.27.1
sniffio==1.3.1
starlette==0.36.3
tomli==2.0.1
twine==5.0.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
virtualenv==20.26.1
wcwidth==0.2.13
yarl==1.9.4
zipp==3.18.1
